Add files via upload
Our project is all about food reviews  to predict the sentiment of a given text using python where we use  Natural Language Processing Toolkit [NLKT], a package in python made especially for text-based analysis. So with a few lines of code, we can easily predict whether a sentence or a review is a positive or a negative review . 
                     Thanks to the ineuron team for contributing with us to learn all these wonder things . We did this project successfully because of the active contribution by the ineuron team and our team work as a part of HACKOTHON December in our KIET college .
OUR TEAM MEMBERS :-1.  CH. JOSHUVA {LEAD}
                                             2.    VAKA . PERI REDDY
                                              3.    G .CHARANTEJA
                                                4.    Y . DHILEEP
                                                 5.        B . SURYA ARJUN